Changing ttir-builder structure
